Fix the bug preventing comments to be entered for possible duplicates detected during issue entry.
* src/java/org/tigris/scarab/tools/ScarabRequestTool.java
  getIssues(): Issue PKs should be of type Long, rather than Integer.


* src/conf/intake.xml
  Issue.Ids: The field should be of type long, rather than int.

Issue: PCN # 17851

+++ src/java/org/tigris/scarab/tools/ScarabRequestTool.java
@@ -1449,8 +1449,7 @@ e.printStackTrace();
             .get("Issue", IntakeTool.DEFAULT_KEY, false);
         if (issueGroup != null) 
         {            
-            Integer[] issueIds =  (Integer[])
-                issueGroup.get("Ids").getValue();
+            Long[] issueIds =  (Long []) issueGroup.get("Ids").getValue();
             if (issueIds != null)
             {
                 issues = getIssues(Arrays.asList(issueIds));
